Refactoring
Moved from GameEngineBuilder to GameState.createGameDefinition, GameDefinition.createNextGamePieceGenerator
