+++ app/src/main/java/de/mwvb/blockpuzzle/game/stonewars/deathstar/DeathStarClassicGameDefinition.java
@@ -2,19 +2,76 @@ package de.mwvb.blockpuzzle.game.stonewars.deathstar;
 
 import java.util.Random;
 
+import de.mwvb.blockpuzzle.block.BlockTypes;
 import de.mwvb.blockpuzzle.game.GameEngineBuilder;
 import de.mwvb.blockpuzzle.gamedefinition.ClassicGameDefinition;
+import de.mwvb.blockpuzzle.gamepiece.GamePiece;
+import de.mwvb.blockpuzzle.gamepiece.INextGamePiece;
+import de.mwvb.blockpuzzle.gamepiece.INextRound;
+import de.mwvb.blockpuzzle.gamepiece.NextGamePieceAdapter;
+import de.mwvb.blockpuzzle.gamepiece.NextGamePieceFromSet;
+import de.mwvb.blockpuzzle.gamestate.GameState;
+import de.mwvb.blockpuzzle.gamestate.Spielstand;
+import de.mwvb.blockpuzzle.gamestate.SpielstandDAO;
+import de.mwvb.blockpuzzle.gamestate.StoneWarsGameState;
 import de.mwvb.blockpuzzle.global.Features;
 import de.mwvb.blockpuzzle.global.messages.MessageFactory;
 import de.mwvb.blockpuzzle.global.messages.MessageObjectWithGameState;
 import de.mwvb.blockpuzzle.playingfield.PlayingField;
 
 public class DeathStarClassicGameDefinition extends ClassicGameDefinition {
+    private static final int NEXTROUND_INDEX = 99; // Ich will eine eigene Datei für den Todesstern-NextRound-Wert.
+    private static final Spielstand nextRoundSS = new Spielstand();
     private boolean won = false; // TO-DO persistieren
 
     public DeathStarClassicGameDefinition(int gamePieceSetNumber, int minimumLiberationScore, int name) {
         super(gamePieceSetNumber, minimumLiberationScore);
         setTerritoryName(name);
+    }
+
+    @Override
+    public INextGamePiece createNextGamePieceGenerator(GameState gs) {
+        // Spielsteine sollen nicht pro Reaktor sein, sondern für Todesstern.
+        final DeathStar planet = (DeathStar) ((StoneWarsGameState) gs).getPlanet();
+        final SpielstandDAO dao = new SpielstandDAO();
+        INextRound persistence = new INextRound() {
+            @Override
+            public void saveNextRound(int val) {
+                writeNextRound(val, dao);
+            }
+
+            @Override
+            public int getNextRound() {
+                return dao.load(planet, NEXTROUND_INDEX).getNextRound();
+            }
+        };
+        NextGamePieceFromSet nextGamePiece = new NextGamePieceFromSet(planet.getGameDefinitions().get(0).getGamePieceSetNumber(), persistence);
+        // Change color of all game pieces. Each reactor has its own color.
+        return new NextGamePieceAdapter(nextGamePiece) {
+            @Override
+            public GamePiece next(BlockTypes blockTypes) {
+                GamePiece gp = super.next(blockTypes);
+                gp.color(getColor());
+                return gp;
+            }
+
+            private int getColor() {
+                switch (planet.getGameIndex()) {
+                    case 0:
+                        if (Features.developerMode) return 6; // yellow
+                        return 11; // dark blue for 1st reactor
+                    case 1:  return  4; // blue for 2nd reactor
+                    default:
+                        if (Features.developerMode) return 3; // red
+                        return  5; // pink for last reactor
+                }
+            }
+        };
+    }
+
+    public void writeNextRound(int val, SpielstandDAO dao) {
+        nextRoundSS.setNextRound(val);
+        dao.save(MilkyWayCluster.INSTANCE.get(), NEXTROUND_INDEX, nextRoundSS);
     }
 
     @Override
